Don't use separate files for item ostructs.
