+++ vnet/lib/vnet.rb
@@ -59,11 +59,11 @@ module Vnet
     autoload :ActiveRouteLinkEvents, 'vnet/core/event_helpers'
     autoload :DpInfo, 'vnet/core/dp_info'
 
-    autoload :ActiveInterface, 'vnet/core/active_interface'
+    autoload :ActiveInterface, 'vnet/core/items'
     autoload :ActiveInterfaceManager, 'vnet/core/active_interface_manager'
     autoload :AddressHelpers, 'vnet/core/address_helpers'
     autoload :ConnectionManager, 'vnet/core/connection_manager'
-    autoload :Datapath, 'vnet/core/datapath'
+    autoload :Datapath, 'vnet/core/items'
     autoload :DatapathManager, 'vnet/core/datapath_manager'
     autoload :FilterManager, 'vnet/core/filter_manager'
     autoload :HostDatapath, 'vnet/core/host_datapath'
@@ -73,17 +73,17 @@ module Vnet
     autoload :InterfacePort, 'vnet/core/interface_port'
     autoload :InterfacePortManager, 'vnet/core/interface_port_manager'
     autoload :Manager, 'vnet/core/manager'
-    autoload :Network, 'vnet/core/network'
+    autoload :Network, 'vnet/core/items'
     autoload :NetworkManager, 'vnet/core/network_manager'
     autoload :Port, 'vnet/core/port'
     autoload :PortManager, 'vnet/core/port_manager'
-    autoload :Route, 'vnet/core/route'
+    autoload :Route, 'vnet/core/items'
     autoload :RouteManager, 'vnet/core/route_manager'
-    autoload :Router, 'vnet/core/router'
+    autoload :Router, 'vnet/core/items'
     autoload :RouterManager, 'vnet/core/router_manager'
     autoload :Service, 'vnet/core/service'
     autoload :ServiceManager, 'vnet/core/service_manager'
-    autoload :Translation, 'vnet/core/translation'
+    autoload :Translation, 'vnet/core/items'
     autoload :TranslationManager, 'vnet/core/translation_manager'
     autoload :Tunnel, 'vnet/core/tunnel'
     autoload :TunnelManager, 'vnet/core/tunnel_manager'
