--- vnet/lib/vnet/core/active_interface.rb
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-require 'ostruct'
-
-module Vnet::Core
-
-  class ActiveInterface < OpenStruct
-  end
-
-end
