--- vnet/lib/vnet/core/datapath.rb
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-require 'ostruct'
-
-module Vnet::Core
-
-  class Datapath < OpenStruct
-  end
-
-end
