+++ vnet/lib/vnet/core/items.rb
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+
+require 'ostruct'
+
+module Vnet::Core
+
+  class ActiveInterface < OpenStruct
+  end
+
+  class Datapath < OpenStruct
+  end
+
+  class Network < OpenStruct
+  end
+
+  class Route < OpenStruct
+  end
+
+  class Router < OpenStruct
+  end
+
+  class Translation < OpenStruct
+  end
+
+end
