--- vnet/lib/vnet/core/network.rb
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-require 'ostruct'
-
-module Vnet::Core
-
-  class Network < OpenStruct
-  end
-
-end
