--- vnet/lib/vnet/core/route.rb
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-require 'ostruct'
-
-module Vnet::Core
-
-  class Route < OpenStruct
-  end
-
-end
