--- vnet/lib/vnet/core/router.rb
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-require 'ostruct'
-
-module Vnet::Core
-
-  class Router < OpenStruct
-  end
-
-end
