--- vnet/lib/vnet/core/translation.rb
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-require 'ostruct'
-
-module Vnet::Core
-
-  class Translation < OpenStruct
-
-  end
-
-end
